fix: moved conversation app credentials related classes to Sinch.Conversation.Apps.Credentials namespace
