+++ src/Sinch/Conversation/Apps/App.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Text.Json.Serialization;
 using Sinch.Conversation.Apps.Create;
+using Sinch.Conversation.Apps.Credentials;
 using Sinch.Conversation.Common;
 
 namespace Sinch.Conversation.Apps
