+++ src/Sinch/Conversation/Apps/Create/CreateAppRequest.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Text.Json.Serialization;
+using Sinch.Conversation.Apps.Credentials;
 using Sinch.Conversation.Common;
 
 namespace Sinch.Conversation.Apps.Create
